transformacao_de_dados/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pandas==2.2.3
tabula-py==2.10.0
